# db: add NextPrefix method to Iterator 