--- a/error_iter.go
+++ b/error_iter.go
@@ -49,6 +49,10 @@ func (c *errorIter) Next() (*InternalKey, []byte) {
 	return nil, nil
 }
 
+func (c *errorIter) NextPrefix(int) (*InternalKey, []byte) {
+	return nil, nil
+}
+
 func (c *errorIter) Prev() (*InternalKey, []byte) {
 	return nil, nil
 }