--- a/iterator_test.go
+++ b/iterator_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/cockroachdb/pebble/internal/base"
 	"github.com/cockroachdb/pebble/internal/datadriven"
 	"github.com/cockroachdb/pebble/internal/manifest"
+	"github.com/cockroachdb/pebble/internal/testkeys"
 	"github.com/cockroachdb/pebble/sstable"
 	"github.com/cockroachdb/pebble/vfs"
 )
@@ -153,6 +154,10 @@ func (f *fakeIter) Next() (*InternalKey, []byte) {
 	}
 	f.valid = true
 	return f.Key(), f.Value()
+}
+
+func (f *fakeIter) NextPrefix(int) (*InternalKey, []byte) {
+	return f.Next()
 }
 
 func (f *fakeIter) Prev() (*InternalKey, []byte) {
@@ -292,6 +297,10 @@ func (i *invalidatingIter) Last() (*InternalKey, []byte) {
 
 func (i *invalidatingIter) Next() (*InternalKey, []byte) {
 	return i.update(i.iter.Next())
+}
+
+func (i *invalidatingIter) NextPrefix(currentPrefixLen int) (*InternalKey, []byte) {
+	return i.update(i.iter.NextPrefix(currentPrefixLen))
 }
 
 func (i *invalidatingIter) Prev() (*InternalKey, []byte) {
@@ -1703,6 +1712,61 @@ func BenchmarkBlockPropertyFilter(b *testing.B) {
 					require.NoError(b, iter.Close())
 				})
 			}
+		})
+	}
+}
+
+func BenchmarkNextPrefix(b *testing.B) {
+	const keyCount = 1000
+	const valueSize = 10
+
+	for _, versionCount := range []int{1, 2, 10, 100, 1000} {
+		b.Run(fmt.Sprintf("versions=%d", versionCount), func(b *testing.B) {
+			r := rand.New(rand.NewSource(7))
+			mem := vfs.NewMem()
+			opts := &Options{
+				Comparer:           testkeys.Comparer,
+				FS:                 mem,
+				FormatMajorVersion: FormatNewest,
+			}
+			d, err := Open("", opts)
+			require.NoError(b, err)
+
+			ks := testkeys.Alpha(8)
+			if v := ks.Count() / keyCount; v > 0 {
+				ks = ks.EveryN(v)
+			}
+			key := make([]byte, ks.MaxLen()+testkeys.MaxSuffixLen)
+			value := make([]byte, valueSize)
+			for i := 0; i < keyCount; i++ {
+				batch := d.NewBatch()
+				n := testkeys.WriteKey(key, ks, i)
+				for j := 0; j < versionCount; j++ {
+					sn := testkeys.WriteSuffix(key[n:], j)
+					r.Read(value)
+					require.NoError(b, batch.Set(key[:n+sn], value, nil))
+				}
+				require.NoError(b, batch.Commit(nil))
+			}
+			require.NoError(b, d.Flush())
+			require.NoError(b, d.Compact(nil, []byte{0xFF}))
+
+			iter := d.NewIter(nil)
+			var prevPrefix []byte
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				count := 0
+				prevPrefix = prevPrefix[:0]
+				valid := iter.First()
+				for valid {
+					count++
+					valid = iter.NextPrefix()
+				}
+				require.Equal(b, keyCount, count)
+			}
+			b.StopTimer()
+			require.NoError(b, iter.Close())
+			require.NoError(b, d.Close())
 		})
 	}
 }