Expose a NextPrefix method on *pebble.Iterator that may be called when the user
wishes to move to the next key prefix (as determined by Comparer.Split).
NextPrefix guarantees that it will move to the next key prefix.

Here's a benchmark comparison of two versions of the benchmark. On the 'old'
benchmark, there is no NextPrefix optimization and the benchmark manually
checks for a new unique prefix. In the 'new' benchmark, the benchmark uses
NextPrefix and omits the manual checks since NextPrefix guarantees the next
returned key will be a new prefix:

```
name                         old time/op    new time/op    delta
NextPrefix/versions=1-24        135µs ± 0%     130µs ± 1%   -4.25%  (p=0.000 n=10+10)
NextPrefix/versions=2-24        313µs ± 1%     154µs ± 0%  -50.97%  (p=0.000 n=10+10)
NextPrefix/versions=10-24      1.71ms ± 0%    0.35ms ± 1%  -79.70%  (p=0.000 n=10+10)
NextPrefix/versions=100-24     12.0ms ± 3%     2.6ms ± 2%  -78.37%  (p=0.000 n=10+9)
NextPrefix/versions=1000-24     142ms ± 1%      53ms ± 1%  -62.72%  (p=0.000 n=10+10)

name                         old alloc/op   new alloc/op   delta
NextPrefix/versions=1-24        10.0B ± 0%     10.0B ± 0%     ~     (all equal)
NextPrefix/versions=2-24        13.0B ± 0%     10.0B ± 0%  -23.08%  (p=0.002 n=8+10)
NextPrefix/versions=10-24       35.4B ± 2%     13.0B ± 0%  -63.28%  (p=0.000 n=10+10)
NextPrefix/versions=100-24       137B ±29%       36B ± 0%  -73.66%  (p=0.000 n=10+8)
NextPrefix/versions=1000-24    3.26kB ± 2%    1.24kB ± 4%  -61.92%  (p=0.000 n=9+8)

name                         old allocs/op  new allocs/op  delta
NextPrefix/versions=1-24         1.00 ± 0%      1.00 ± 0%     ~     (all equal)
NextPrefix/versions=2-24         1.00 ± 0%      1.00 ± 0%     ~     (all equal)
NextPrefix/versions=10-24        1.00 ± 0%      1.00 ± 0%     ~     (all equal)
NextPrefix/versions=100-24       1.40 ±43%      1.00 ± 0%     ~     (p=0.087 n=10+10)
NextPrefix/versions=1000-24      28.4 ± 5%      11.0 ± 0%  -61.23%  (p=0.000 n=8+8)
```