@jbowens 2022-01-15
Metamorphic tests revealed an issue where we don't always advance to the next prefix. I'll look into it.